Refactor (#30)
* Refactor backend

* Update CI and Makefile

* Fix dev-server command

* Fix uv sync

---------
