+++ apps/backend/pyproject.toml
@@ -1,27 +1,30 @@
 [build-system]
-requires = ["setuptools>=67"]
 build-backend = "setuptools.build_meta"
+requires = ["setuptools>=67"]
 
 [project]
-name = "orcheo-backend"
-version = "0.1.0"
+dependencies = [
+  "fastapi>=0.104.0",
+  "uvicorn[standard]>=0.24.0",
+  "orcheo>=0.0.1"
+]
 description = "Deployment wrapper around the Orcheo FastAPI application"
+name = "orcheo-backend"
 readme = "README.md"
 requires-python = ">=3.12"
-dependencies = [
-    "fastapi>=0.104.0",
-    "uvicorn[standard]>=0.24.0",
-    "orcheo>=0.0.1"
-]
+version = "0.0.1"
 
 [project.optional-dependencies]
 dev = [
-    "pytest",
-    "pytest-asyncio",
-    "pytest-cov",
-    "ruff",
-    "mypy"
+  "pytest",
+  "pytest-asyncio",
+  "pytest-cov",
+  "ruff",
+  "mypy"
 ]
 
 [project.scripts]
 orcheo-backend = "orcheo_backend:create_app"
+
+[tool.uv.sources]
+orcheo = {workspace = true}
